python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=./python/requirements.txt ./python/setup.py
#
aiohttp==3.8.5
    # via openai
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
blinker==1.6.2
    # via flask
brotli==1.1.0
    # via geventhttpclient
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via
    #   geventhttpclient
    #   kubernetes
    #   requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via flask
configargparse==1.7
    # via locust
flask==2.3.3
    # via
    #   flask-basicauth
    #   flask-cors
    #   locust
flask-basicauth==0.2.0
    # via locust
flask-cors==4.0.0
    # via locust
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
gevent==23.9.0.post1
    # via
    #   geventhttpclient
    #   locust
geventhttpclient==2.0.10
    # via locust
google-auth==2.22.0
    # via kubernetes
greenlet==2.0.2
    # via gevent
idna==3.4
    # via
    #   requests
    #   yarl
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
kubernetes==27.2.0
    # via AutoChaos (./python/setup.py)
locust==2.16.1
    # via AutoChaos (./python/setup.py)
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
msgpack==1.0.5
    # via locust
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
openai==0.28.0
    # via AutoChaos (./python/setup.py)
psutil==5.9.5
    # via locust
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
python-dateutil==2.8.2
    # via kubernetes
python-dotenv==1.0.0
    # via AutoChaos (./python/setup.py)
pyyaml==6.0.1
    # via kubernetes
pyzmq==25.1.1
    # via locust
requests==2.31.0
    # via
    #   kubernetes
    #   locust
    #   openai
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
roundrobin==0.0.4
    # via locust
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   geventhttpclient
    #   google-auth
    #   kubernetes
    #   python-dateutil
tqdm==4.66.1
    # via openai
typing-extensions==4.7.1
    # via locust
urllib3==1.26.16
    # via
    #   google-auth
    #   kubernetes
    #   requests
websocket-client==1.6.2
    # via kubernetes
werkzeug==2.3.7
    # via
    #   flask
    #   locust
yarl==1.9.2
    # via aiohttp
zope-event==5.0
    # via gevent
zope-interface==6.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
